cookies/secure/set-from-dom.https.sub.html

<!doctype html>
<html>
<head>
  <meta charset=utf-8>
  <title>Set 'secure' cookie from `document.cookie` on a secure page</title>
  <meta name=help href="https://tools.ietf.org/html/draft-west-leave-secure-cookies-alone">
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/cookies/resources/testharness-helpers.js"></script>
</head>
<body>
<div id=log></div>
<script>
    function build_arguments_str(args, separator) {
        return args[0] + " " + separator + " " + args[1];
    }
    function build_arguments_str_fetch(args, separator) {
        return args[0] + " " + separator + " " + JSON.stringify(args[1]);
    }
    function setup_logs(){
        // hook the document.cookie setters and getters
        var cookie_setter_orig = document.__lookupSetter__("cookie").bind(document);
        var cookie_getter_orig = document.__lookupGetter__("cookie").bind(document);
        Object.defineProperty(document, "cookie", {
        get: function () {
            let a = cookie_getter_orig();
            console.log("[{{host}}]: [document.cookie] GET -> " + a);
            return a;
        },
        set: function (val) {
            console.log("[{{host}}]: [document.cookie] SET -> " + val);
            cookie_setter_orig(val);
        }
        });

        // hook the assert
        assert_equals = new Proxy(assert_equals, {
            apply:
                function(target, thisArg, args) {
                    console.log("[{{host}}]: [assert_equals] -> " + build_arguments_str(args, "==") );
                    console.log("======== TEST ========");
                    return target.apply(thisArg, args);
                }
        });
        assert_not_equals = new Proxy(assert_not_equals, {
            apply:
                function(target, thisArg, args) {
                    console.log("[{{host}}]: [assert_not_equals] -> " + build_arguments_str(args, "!=") );
                    console.log("======== TEST ========");
                    return target.apply(thisArg, args);
                }
        });

        fetch = new Proxy(fetch, {
            apply:
                function(target, thisArg, args) {
                    console.log("[fetch] -> " + build_arguments_str_fetch(args, "|") );
                    return target.apply(thisArg, args);
                }
        });
    }
  setup_logs();
  var tests = [
    [
      "'secure' cookie visible in `document.cookie`",
      function () {
        document.cookie = "secure_from_secure_dom=1; secure; path=/";
        assert_not_equals(document.cookie.match(/secure_from_secure_dom=1/), null);
        this.done();
      }
    ],
    [
      "'secure' cookie visible in HTTP request",
      function () {
        document.cookie = "secure_from_secure_dom=1; secure; path=/";
        assert_not_equals(document.cookie.match(/secure_from_secure_dom=1/), null);
        fetch("https://{{host}}:{{ports[https][0]}}/cookies/resources/echo-json.py",
              { "credentials": "include" })
          .then(this.step_func(function (r) {
            return r.json();
          }))
          .then(this.step_func_done(function (j) {
            assert_equals(j["secure_from_secure_dom"], "secure_from_secure_dom=1");
          }));
      }
    ]
  ];

  function clearKnownCookie() {
    document.cookie = "secure_from_secure_dom=0; Secure; expires=Thu, 01 Jan 1970 00:00:01 GMT; path=/";
  }

  executeTestsSerially(tests, clearKnownCookie, clearKnownCookie);
</script>
</body>
</html>
